chore: reorganzing ServicesTable file structure
